DEV-2305 Use dest path to construct target folder
The target folder ENV var contained the absolute path of the folder to
unzip the zip file into. Depending on the sipin flow that path can be
different. An option could be making multiple ENV vars per flow. However
the destination path of the incoming event is also an absolute path
pointing to the transferred zip file. In practice only the last
subfolder of both folders will be different.

Example:
 - Destination path: /path/to/sipin/incoming/file.zip
 - Target unzip folder: /path/to/sipin/unzip/file.zip/

So, reuse the destination path to construct the target folder. The target
folder ENV var is then used to define the different subfolder.

Also, fix some flake8 warnings.